services/results_adapter/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
prometheus-client==0.20.0
requests==2.32.3
paho-mqtt==2.1.0
pynacl==1.5.0

